"Validation" for preventing duplicate dates on DateForm
This validation checks the db, pulls out all dateslot models, and raises a ValidationError if the date of any model matches the date of the one we're validating.
The problem with this is that it will check all models across all events, and we want it only to check ones with the same part_id. However I don't know of a way to know the part_id during validation, since the DateForm class is not tied to the part_id.